Move Schematic's rendering related methods to the separate file
